pyproject.toml
[project]
name = "em-backend"
version = "0.1.0"
description = "The Backend for Electomate"
readme = "README.md"
authors = [
    { name = "BrokenDuck", email = "jonathan.maillefaud@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "aiostream>=0.7.0",
    "alembic>=1.16.5",
    "asyncpg>=0.29.0",
    "alembic-postgresql-enum>=1.8.0",
    "alembic-utils>=0.8.8",
    "asgi-lifespan>=2.1.0",
    "azure-monitor-opentelemetry>=1.8.1",
    "docling>=2.54.0",
    "docling-core[chunking-openai]>=2.48.2",
    "fastapi[standard-no-fastapi-cloud-cli]>=0.116.2",
    "httpx>=0.28.1",
    "langchain-core>=0.3.76",
    "langchain-openai>=0.3.33",
    "langchain-text-splitters>=0.3.11",
    "langgraph>=0.6.7",
    "openai>=1.108.0",
    "opentelemetry-api>=1.37.0",
    "opentelemetry-instrumentation-fastapi>=0.58b0",
    "opentelemetry-instrumentation-httpx>=0.58b0",
    "opentelemetry-sdk>=1.37.0",
    "orjson>=3.11.3",
    "pydantic-settings>=2.10.1",
    "python-json-logger>=3.3.0",
    "pymupdf>=1.24.10",
    "sqlalchemy[postgresql-psycopgbinary]>=2.0.43",
    "starlette-context>=0.4.0",
    "structlog>=25.4.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "truststore>=0.10.4",
    "weaviate-client>=4.16.10",
    "pypdf>=6.1.3",
    # Knowledge Graph dependencies
    "psycopg2-binary>=2.9.9",
    "google-genai>=1.0.0",
    "yt-dlp>=2024.0.0",
    "playwright>=1.40.0",
    "pdfplumber>=0.11.0",
    "lxml>=5.0.0",
]

[dependency-groups]
dev = []


[project.scripts]
em-backend = "em_backend:main"

[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu" },
]
torchvision = [
  { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff.lint]
select = [
    # FastAPI
    "FAST",
    # flake8-annotations: detect missing annotations
    "ANN",
    # flake8-async: detect async problems
    "ASYNC",
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]


[tool.alembic]

# path to migration scripts.
# this is typically a path given in POSIX (e.g. forward slashes)
# format, relative to the token %(here)s which refers to the location of this
# ini file
script_location = "%(here)s/alembic"

# template used to generate migration file names; The default value is %%(rev)s_%%(slug)s
# Uncomment the line below if you want the files to be prepended with date and time
# see https://alembic.sqlalchemy.org/en/latest/tutorial.html#editing-the-ini-file
# for all available tokens
# file_template = "%%(year)d_%%(month).2d_%%(day).2d_%%(hour).2d%%(minute).2d-%%(rev)s_%%(slug)s"

# additional paths to be prepended to sys.path. defaults to the current working directory.
prepend_sys_path = [
    "."
]


# timezone to use when rendering the date within the migration file
# as well as the filename.
# If specified, requires the python>=3.9 or backports.zoneinfo library and tzdata library.
# Any required deps can installed by adding `alembic[tz]` to the pip requirements
# string value is passed to ZoneInfo()
# leave blank for localtime
# timezone =

# max length of characters to apply to the "slug" field
# truncate_slug_length = 40

# set to 'true' to run the environment during
# the 'revision' command, regardless of autogenerate
# revision_environment = false

# set to 'true' to allow .pyc and .pyo files without
# a source .py file to be detected as revisions in the
# versions/ directory
# sourceless = false

# version location specification; This defaults
# to <script_location>/versions.  When using multiple version
# directories, initial revisions must be specified with --version-path.
# version_locations = [
#    "%(here)s/alembic/versions",
#    "%(here)s/foo/bar"
# ]


# set to 'true' to search source files recursively
# in each "version_locations" directory
# new in Alembic version 1.10
# recursive_version_locations = false

# the output encoding used when revision files
# are written from script.py.mako
# output_encoding = "utf-8"

# This section defines scripts or Python functions that are run
# on newly generated revision scripts.  See the documentation for further
# detail and examples
# [[tool.alembic.post_write_hooks]]
# format using "black" - use the console_scripts runner,
# against the "black" entrypoint
# name = "black"
# type = "console_scripts"
# entrypoint = "black"
# options = "-l 79 REVISION_SCRIPT_FILENAME"
#
# [[tool.alembic.post_write_hooks]]
# lint with attempts to fix using "ruff" - use the module runner, against the "ruff" module
# name = "ruff"
# type = "module"
# module = "ruff"
# options = "check --fix REVISION_SCRIPT_FILENAME"
#
# [[tool.alembic.post_write_hooks]]
# Alternatively, use the exec runner to execute a binary found on your PATH
# name = "ruff"
# type = "exec"
# executable = "ruff"
# options = "check --fix REVISION_SCRIPT_FILENAME"

